Latest daily variation of Morrocan stocks

Get the last variation of all stocks that are available in the stock exchange and the MASI/MADEX indices.

get_today()

Value

Returns a data.frame containing the daily variation for all the stocks.

name: name of the company/stock.

price: last price.

variation: variation of the price.

Examples

get_today()

#>                   name   price variation
#> 1               Addoha    6.01     -2.28
#> 2                 AFMA 1039.00     -0.10
#> 3         Afric Indus.  267.20      1.98
#> 4         Afriquia Gaz 3920.00      0.00
#> 5                 Agma 3364.00      0.00
#> 6            Alliances   37.48      0.64
#> 7      Aluminium Maroc 1299.00      0.08
#> 8         ATLANTASANAD   81.95      1.17
#> 9   Attijariwafa Bk Rg  409.95      0.72
#> 11           Auto Hall   70.00     -0.92
#> 12          Auto Nejma 1978.00      0.00
#> 13              BALIMA   96.45      0.00
#> 14              BoA Br  151.45     -0.75
#> 15                 BCP  245.00      0.20
#> 17                BMCI  611.00      0.00
#> 18       Cartier Saada   28.00      0.00
#> 19                 CDM  443.00      0.00
#> 20      Central.Danone  523.90      0.00
#> 21                 CIH  252.00     -1.18
#> 22       Ciments Maroc 1544.00      0.59
#> 23                 CMT 1490.00      0.34
#> 24            Colorado   47.70      0.40
#> 25             COSUMAR  210.00     -0.47
#> 26                 CTM  680.70     -3.99
#> 27       Dari Couspate 4700.00      0.00
#> 28       Delattre Lev.   33.36     -3.97
#> 29       Delta Holding   31.97     -0.03
#> 30          Diac Salaf   26.25      0.00
#> 31              DISWAY  438.50      0.11
#> 32              Ennakl   25.40     -1.17
#> 33               EQDOM 1080.00      0.00
#> 34     FENIE BROSSETTE   41.29      1.98
#> 35                 HPS 5800.00     -0.85
#> 36         IBMaroc.com   29.00      0.35
#> 37      Immr Invest Br   92.25     -2.38
#> 38             INVOLYS  117.00      3.54
#> 39     Jet Contractors  185.00      1.09
#> 40           LABEL VIE 3100.00      1.27
#> 41 LafargeHolcim Ma Br 1550.00     -0.32
#> 42     Lesieur Cristal  162.00      2.53
#> 43               Lydec  364.95      3.68
#> 44           M2M Group  640.00      0.00
#> 45     Maghreb Oxygene  186.30      0.00
#> 46          Maghrebail  873.00      0.00
#> 47             Managem 1098.00      0.18
#> 48       Maroc Leasing  479.90      0.00
#> 49       Maroc Telecom  144.00      0.00
#> 50           Med Paper   14.11      0.71
#> 51           Microdata  569.00      2.52
#> 52         Mutandis Br  220.00      0.00
#> 53        Nexans Maroc  130.00      0.00
#> 54              Oulmes 1297.00      0.00
#> 55          PROMOPHARM  715.00      0.00
#> 56       Rebab Company  109.95     -1.48
#> 57    Res.Dar Saada Br   25.10     -0.79
#> 58               Risma  105.00      0.00
#> 59                 S2M  166.00      3.69
#> 60     Saham Assurance 1180.00      1.72
#> 61             SALAFIN  530.00      0.00
#> 62               SAMIR  133.50      0.00
#> 63                 SMI 1900.00     -0.94
#> 64   Stokvis Nord Afr.   10.89     -3.11
#> 65                SNEP  429.95     -0.01
#> 66             Sonasid  270.35      3.98
#> 67             SOTHEMA 1708.00      0.00
#> 68                 SRM  142.35      0.00
#> 69        Ste Boissons 2300.00     -2.25
#> 70        STROC Indus.   21.94     -0.05
#> 71        TAQA Morocco  936.00     -1.68
#> 72               Timar  127.45      0.00
#> 73         Total Maroc 1050.00      0.86
#> 74              Unimer  164.00      0.00
#> 75          Wafa Assur 3551.00     -1.36
#> 76            Zellidja   76.13      0.00