Intraday Prices

Scraps the intraday stock prices for the provided companies

intraday(name)

intradays(...)

Arguments

name, ...
Company names as returned by the get_today() function

Value

Returns a data.frame containing all the intraday prices for the specified stocks.

time: the time of the record.

price: the price at that time. Where the symbol column contains the name of the company.

Examples

if( bcr:::check_internet() ) {
head(intraday(), 30)
}

#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       time                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    price
#> 1 09:30:17, 09:30:18, 09:30:22, 09:30:30, 09:30:35, 09:30:37, 09:30:38, 09:30:40, 09:30:41, 09:30:44, 09:30:45, 09:30:50, 09:30:51, 09:30:59, 09:31:39, 09:34:37, 09:34:37, 09:35:05, 09:36:15, 09:36:15, 09:36:25, 09:40:44, 09:41:03, 09:41:03, 09:41:21, 09:41:22, 09:41:22, 09:41:22, 09:43:10, 09:43:10, 09:44:11, 09:44:11, 09:45:02, 09:45:06, 09:45:06, 09:45:06, 09:45:06, 09:46:32, 09:46:32, 09:46:32, 09:46:36, 09:46:36, 09:47:32, 09:47:32, 09:48:12, 09:48:29, 09:50:36, 09:51:22, 09:53:01, 09:54:43, 09:56:00, 09:56:00, 09:56:00, 09:56:00, 09:58:54, 09:59:12, 09:59:12, 10:00:50, 10:01:14, 10:05:12, 10:10:30, 10:14:18, 10:15:43, 10:15:43, 10:20:16, 10:20:16, 10:20:16, 10:22:47, 10:24:38, 10:24:51, 10:27:36, 10:27:36, 10:27:37, 10:27:37, 10:27:37, 10:27:37, 10:27:37, 10:27:37, 10:27:37, 10:27:37, 10:29:30, 10:30:22, 10:30:22, 10:30:22, 10:30:22, 10:34:09, 10:36:40, 10:39:04, 10:48:05, 10:53:20, 10:53:41, 10:53:41, 10:55:08, 10:55:36, 11:06:46, 11:06:46, 11:06:46, 11:06:46, 11:06:46, 11:06:46, 11:07:50, 11:11:39, 11:12:05, 11:15:45, 11:16:04, 11:16:12, 11:16:45, 11:19:04, 11:20:36, 11:20:36, 11:21:27, 11:22:27, 11:22:27, 11:22:27, 11:25:13, 11:27:41, 11:27:41, 11:27:41, 11:29:04, 11:29:35, 11:29:35, 11:29:35, 11:29:51, 11:29:51, 11:33:48, 11:34:30, 11:36:29, 11:38:32, 11:41:45, 11:41:45, 11:45:18, 11:45:24, 11:46:54, 11:48:14, 11:48:54, 11:50:10, 11:54:16, 11:54:23, 11:55:28, 11:55:28, 11:55:28, 11:58:49, 12:00:15, 12:00:15, 12:00:15, 12:03:02, 12:07:13, 12:07:13, 12:07:13, 12:07:13, 12:07:43, 12:16:05, 12:16:52, 12:16:52, 12:17:05, 12:17:08, 12:17:08, 12:17:08, 12:17:24, 12:18:10, 12:22:28, 12:23:20, 12:32:41, 12:33:37, 12:37:26, 12:38:26, 12:38:26, 12:43:25, 12:45:10, 12:47:50, 12:49:11, 12:49:16, 12:49:16, 12:49:18, 12:50:15, 12:50:15, 12:50:32, 12:50:41, 12:50:46, 12:51:11, 12:51:11, 12:51:11, 12:52:42, 12:53:15, 12:53:15, 12:54:11, 12:58:39, 13:00:33, 13:03:09, 13:03:46, 13:04:56, 13:11:37, 13:11:37, 13:11:37, 13:11:37, 13:13:07, 13:21:01, 13:21:01, 13:21:18, 13:21:40, 13:21:40, 13:23:30, 13:23:30, 13:25:27, 13:25:33, 13:37:40, 13:38:08, 13:41:24, 13:42:13, 13:42:13, 13:43:44, 13:54:02, 13:55:30, 13:55:30, 13:55:30, 13:55:30, 13:55:30, 13:55:30, 13:55:31, 13:56:54, 13:59:53, 14:03:12, 14:03:26, 14:04:35, 14:04:36, 14:04:47, 14:04:47, 14:04:48, 14:05:15, 14:05:55, 14:06:52, 14:06:53, 14:12:08 12466.69, 12466.72, 12466.20, 12466.29, 12466.31, 12479.18, 12478.95, 12477.92, 12473.93, 12473.36, 12473.36, 12473.75, 12476.06, 12476.27, 12476.34, 12476.35, 12476.39, 12476.27, 12476.27, 12476.27, 12476.33, 12477.89, 12477.48, 12477.47, 12477.20, 12477.21, 12477.43, 12477.47, 12477.37, 12477.47, 12477.48, 12477.48, 12477.48, 12476.59, 12476.55, 12476.49, 12476.45, 12476.82, 12476.89, 12476.90, 12476.89, 12476.90, 12476.90, 12476.90, 12477.12, 12474.71, 12474.60, 12474.68, 12475.70, 12475.41, 12476.20, 12476.32, 12476.43, 12476.54, 12476.91, 12476.50, 12475.23, 12474.87, 12475.23, 12475.15, 12475.03, 12474.67, 12476.39, 12476.10, 12476.38, 12476.38, 12476.39, 12476.39, 12476.87, 12474.01, 12473.94, 12473.93, 12473.85, 12473.85, 12473.79, 12473.78, 12473.77, 12473.77, 12473.69, 12473.65, 12473.67, 12473.70, 12473.67, 12473.67, 12473.63, 12473.63, 12474.01, 12473.63, 12470.91, 12470.62, 12470.66, 12470.62, 12470.66, 12473.52, 12473.37, 12473.31, 12473.30, 12473.24, 12473.23, 12473.23, 12473.12, 12473.04, 12473.05, 12473.17, 12478.74, 12478.89, 12478.89, 12481.75, 12484.61, 12490.05, 12489.93, 12474.49, 12473.63, 12473.20, 12478.78, 12478.59, 12478.68, 12478.78, 12489.26, 12489.74, 12489.97, 12490.21, 12491.06, 12490.64, 12490.29, 12490.60, 12490.70, 12490.70, 12492.16, 12492.27, 12492.36, 12493.03, 12492.83, 12492.93, 12492.04, 12492.24, 12494.08, 12499.80, 12499.86, 12499.80, 12499.75, 12500.43, 12501.66, 12501.68, 12501.70, 12505.99, 12505.69, 12505.68, 12505.68, 12505.67, 12505.75, 12506.42, 12506.11, 12506.08, 12506.50, 12506.20, 12506.12, 12506.10, 12506.02, 12506.13, 12506.61, 12508.98, 12509.12, 12509.78, 12509.78, 12518.29, 12519.35, 12519.79, 12521.12, 12511.59, 12512.48, 12510.17, 12510.13, 12509.86, 12510.21, 12509.89, 12511.53, 12511.54, 12510.21, 12519.74, 12521.88, 12524.50, 12523.82, 12521.88, 12521.88, 12522.11, 12522.33, 12523.66, 12523.55, 12523.56, 12523.56, 12523.57, 12523.56, 12523.52, 12522.92, 12522.85, 12522.78, 12522.70, 12522.96, 12522.96, 12522.96, 12522.32, 12522.27, 12522.51, 12523.17, 12523.18, 12523.06, 12511.63, 12516.22, 12516.54, 12521.38, 12532.81, 12532.88, 12532.84, 12532.78, 12532.67, 12532.64, 12532.57, 12532.50, 12532.51, 12542.09, 12542.08, 12542.09, 12542.08, 12542.07, 12542.65, 12542.68, 12542.44, 12544.52, 12544.54, 12544.55, 12544.58, 12544.60

if( bcr:::check_internet() ) {
intradays("Wafa Assur", "Nexans Maroc")
}

#>                                     time                              price       symbol
#> 1                                   NULL                               NULL Nexans Maroc
#> 2 13:25:31, 13:25:33, 13:25:34, 13:26:31 4000.00, 4010.00, 4010.00, 4010.00   Wafa Assur