Intraday Prices

Scraps the intraday stock prices for the provided companies

intraday(name)

intradays(...)

Arguments

name, ...
Company names as returned by the get_today() function

Value

Returns a data.frame containing all the intraday prices for the specified stocks.

time: the time of the record.

price: the price at that time. Where the symbol column contains the name of the company.

Examples

head(intraday(), 30)

#>        time    price
#> 1  09:30:01 10972.82
#> 2  09:30:03 10973.14
#> 3  09:30:08 10969.61
#> 4  09:30:15 10968.53
#> 5  09:30:27 10968.90
#> 6  09:30:27 10968.37
#> 7  09:30:27 10969.94
#> 8  09:30:36 10969.98
#> 9  09:30:39 10968.69
#> 10 09:30:39 10968.19
#> 11 09:30:41 10968.85
#> 12 09:30:44 10972.75
#> 13 09:30:46 10974.46
#> 14 09:30:47 10973.53
#> 15 09:30:51 10977.54
#> 16 09:30:52 10980.40
#> 17 09:30:56 10994.77
#> 18 09:30:56 10997.08
#> 19 09:30:58 10997.56
#> 20 09:30:58 10997.64
#> 21 09:30:58 10997.71
#> 22 09:31:16 10997.85
#> 23 09:31:16 10997.85
#> 24 09:31:16 10997.86
#> 25 09:31:59 10997.90
#> 26 09:31:59 10997.93
#> 27 09:33:57 10987.91
#> 28 09:34:35 10987.85
#> 29 09:35:32 10987.85
#> 30 09:35:32 10987.85

intradays("Wafa Assur", "Nexans Maroc")

#>       time   price       symbol
#> 1 14:30:31  130.00 Nexans Maroc
#> 2 15:30:52 3551.00   Wafa Assur