Latest daily variation of Moroccan stocks

Get the last variation of all stocks that are available in the stock exchange and the MASI/MADEX indices.

get_today()

Value

Returns a data.frame containing the daily variation for all the stocks.

name: name of the company/stock.

price: last price.

variation: variation of the price.

Examples

if( bcr:::check_internet() ) {
get_today()
}

#>                   name   price variation
#> 1               Addoha   10.21      0.00
#> 2                 AFMA 1220.00      0.00
#> 3         Afric Indus.  305.80      0.00
#> 4         Afriquia Gaz 4189.00      0.00
#> 5                 Agma 4522.00      0.00
#> 6            Alliances   40.65      0.00
#> 7      Aluminium Maroc 1490.00      0.00
#> 8    Aradei Capital Br  392.50      0.00
#> 9         ATLANTASANAD   93.13     -1.40
#> 10  Attijariwafa Bk Rg  466.00      0.73
#> 12           Auto Hall   92.42     -1.99
#> 13          Auto Nejma 2023.00      0.00
#> 14              BALIMA   83.43      0.00
#> 15              BoA Br  160.05      0.00
#> 16                 BCP  267.00      0.00
#> 18                BMCI  633.00     -1.86
#> 19       Cartier Saada   28.20      0.00
#> 20                 CDM  450.00      0.00
#> 21      Central.Danone  558.00      0.00
#> 22                 CIH  272.05     -1.43
#> 23       Ciments Maroc 1900.00      1.06
#> 24                 CMT 1744.00      0.00
#> 25            Colorado   57.00     -2.40
#> 26             COSUMAR  293.50      3.71
#> 27                 CTM  714.00      0.00
#> 28       Dari Couspate 4550.00      0.00
#> 29       Delattre Lev.   39.25      3.29
#> 30       Delta Holding   33.67      0.00
#> 31          Diac Salaf   26.25      0.00
#> 32              DISWAY  570.00      0.00
#> 33              Ennakl   31.00      0.49
#> 34               EQDOM 1260.00      0.00
#> 35     FENIE BROSSETTE   64.45      3.99
#> 36                 HPS 6665.00      0.23
#> 37         IBMaroc.com   30.75      3.26
#> 38      Immr Invest Br   95.07     -0.45
#> 39             INVOLYS  108.00      0.00
#> 40     Jet Contractors  260.00      0.00
#> 41           LABEL VIE 4000.00      1.32
#> 42 LafargeHolcim Ma Br 1858.00      1.98
#> 43     Lesieur Cristal  192.00      1.05
#> 44               Lydec  321.10      0.34
#> 45           M2M Group  860.00      0.00
#> 46     Maghreb Oxygene  190.00      0.00
#> 47          Maghrebail  871.00      0.00
#> 48             Managem 1500.00      1.01
#> 49       Maroc Leasing  431.25      0.00
#> 50       Maroc Telecom  139.00     -0.04
#> 51           Med Paper   15.00      2.32
#> 52           Microdata  653.00     -1.06
#> 53         Mutandis Br  290.00      3.76
#> 54        Nexans Maroc  147.85      0.00
#> 55              Oulmes 1169.00      0.00
#> 56          PROMOPHARM  866.90      0.00
#> 57       Rebab Company  106.00      0.00
#> 58    Res.Dar Saada Br   34.21      0.00
#> 59               Risma  116.50      0.00
#> 60                 S2M  216.00      0.00
#> 61     Saham Assurance 1229.00      0.00
#> 62             SALAFIN  703.00      3.38
#> 63               SAMIR  133.50      0.00
#> 64                 SMI 2139.00      0.00
#> 65   Stokvis Nord Afr.   14.00      3.93
#> 66                SNEP  734.00      0.84
#> 67             Sonasid  553.50      3.98
#> 68             SOTHEMA 4244.00      0.00
#> 69                 SRM  150.00      0.00
#> 70        Ste Boissons 2610.00      0.00
#> 71        STROC Indus.   17.01      0.00
#> 72        TAQA Morocco 1040.00      0.00
#> 73               Timar  131.55      0.00
#> 74         Total Maroc 1470.00     -1.34
#> 75              Unimer  160.00      0.00
#> 76          Wafa Assur 4010.00      0.25
#> 77            Zellidja   75.00      0.00