Latest daily variation of Moroccan stocks

Get the last variation of all stocks that are available in the stock exchange and the MASI/MADEX indices.

get_today()

Value

Returns a data.frame containing the daily variation for all the stocks.

name: name of the company/stock.

price: last price.

variation: variation of the price.

Examples

get_today()

#>                   name   price variation
#> 1               Addoha    6.70     -0.15
#> 2                 AFMA 1110.00      3.54
#> 3         Afric Indus.  277.00      0.00
#> 4         Afriquia Gaz 4000.00     -2.44
#> 5                 Agma 3711.00      0.03
#> 6            Alliances   37.31     -1.82
#> 7      Aluminium Maroc 1325.00      0.00
#> 8    Aradei Capital Br  405.00      0.25
#> 9         ATLANTASANAD   82.90     -0.12
#> 10  Attijariwafa Bk Rg  437.00     -0.67
#> 12           Auto Hall   70.20      1.92
#> 13          Auto Nejma 1985.00      0.00
#> 14              BALIMA   96.45      0.00
#> 15              BoA Br  158.90      2.15
#> 16                 BCP  254.60     -1.70
#> 18                BMCI  645.00      0.00
#> 19       Cartier Saada   29.15      0.00
#> 20                 CDM  465.00      0.22
#> 21      Central.Danone  513.00     -1.91
#> 22                 CIH  270.00      1.89
#> 23       Ciments Maroc 1655.00      1.41
#> 24                 CMT 1530.00     -0.65
#> 25            Colorado   48.50      1.57
#> 26             COSUMAR  221.00     -3.07
#> 27                 CTM  703.00      0.43
#> 28       Dari Couspate 4700.00      0.00
#> 29       Delattre Lev.   35.50     -3.66
#> 30       Delta Holding   31.80      0.00
#> 31          Diac Salaf   26.25      0.00
#> 32              DISWAY  437.00     -0.68
#> 33              Ennakl   27.00     -2.88
#> 34               EQDOM 1120.00     -0.27
#> 35     FENIE BROSSETTE   41.61      0.00
#> 36                 HPS 6595.00     -0.08
#> 37         IBMaroc.com   28.00     -2.91
#> 38      Immr Invest Br   96.37     -0.07
#> 39             INVOLYS  121.45      0.00
#> 40     Jet Contractors  189.00     -0.47
#> 41           LABEL VIE 3290.00      2.17
#> 42 LafargeHolcim Ma Br 1650.00      0.61
#> 43     Lesieur Cristal  164.35      0.00
#> 44               Lydec  370.00      1.66
#> 45           M2M Group  678.00      0.00
#> 46     Maghreb Oxygene  187.00      0.38
#> 47          Maghrebail  873.00      0.00
#> 48             Managem 1140.00     -3.14
#> 49       Maroc Leasing  470.35     -1.99
#> 50       Maroc Telecom  147.00      0.17
#> 51           Med Paper   14.32     -1.85
#> 52           Microdata  596.00      3.47
#> 53         Mutandis Br  221.90      1.32
#> 54        Nexans Maroc  134.95      0.00
#> 55              Oulmes 1242.00     -1.90
#> 56          PROMOPHARM  705.00      0.00
#> 57       Rebab Company  109.95      0.00
#> 58    Res.Dar Saada Br   26.71     -1.91
#> 59               Risma  115.70     -1.11
#> 60                 S2M  166.00      3.69
#> 61     Saham Assurance 1240.00      0.81
#> 62             SALAFIN  531.00      1.72
#> 63               SAMIR  133.50      0.00
#> 64                 SMI 1991.00      2.00
#> 65   Stokvis Nord Afr.   10.83     -0.64
#> 66                SNEP  462.00     -0.75
#> 67             Sonasid  290.50     -2.19
#> 68             SOTHEMA 1937.00      0.00
#> 69                 SRM  135.00     -0.92
#> 70        Ste Boissons 2400.00     -0.25
#> 71        STROC Indus.   20.90      0.00
#> 72        TAQA Morocco  930.00      0.00
#> 73               Timar  127.45      0.00
#> 74         Total Maroc 1218.00     -3.94
#> 75              Unimer  161.00      0.00
#> 76          Wafa Assur 3910.00     -2.25
#> 77            Zellidja   73.12      0.00