Intraday Prices

Scraps the intraday stock prices for the provided companies

intraday(name)

intradays(...)

Arguments

name, ...
Company names as returned by the get_today() function

Value

Returns a data.frame containing all the intraday prices for the specified stocks.

time: the time of the record.

price: the price at that time. Where the symbol column contains the name of the company.

Examples

head(intraday(), 30)

#>        time    price
#> 1  09:30:03 11551.26
#> 2  09:30:03 11551.41
#> 3  09:30:07 11551.56
#> 4  09:30:17 11551.75
#> 5  09:30:17 11554.02
#> 6  09:30:20 11554.13
#> 7  09:30:22 11559.45
#> 8  09:30:28 11558.82
#> 9  09:30:30 11561.09
#> 10 09:30:41 11561.10
#> 11 09:30:41 11561.06
#> 12 09:30:48 11561.29
#> 13 09:30:48 11561.49
#> 14 09:30:49 11566.50
#> 15 09:30:51 11566.83
#> 16 09:30:53 11566.84
#> 17 09:30:53 11555.87
#> 18 09:30:54 11562.64
#> 19 09:30:56 11562.77
#> 20 09:30:57 11562.81
#> 21 09:30:57 11575.89
#> 22 09:31:40 11579.22
#> 23 09:32:00 11579.49
#> 24 09:32:11 11579.59
#> 25 09:32:13 11579.47
#> 26 09:32:27 11579.83
#> 27 09:32:30 11579.75
#> 28 09:32:50 11579.68
#> 29 09:34:03 11579.00
#> 30 09:34:06 11579.12

intradays("Wafa Assur", "Nexans Maroc")

#>       time   price       symbol
#> 1 14:30:30  134.95 Nexans Maroc
#> 2 12:21:21 3995.00   Wafa Assur
#> 3 12:21:21 4000.00   Wafa Assur
#> 4 12:21:21 4050.00   Wafa Assur
#> 5 15:30:00 3910.00   Wafa Assur