Intraday Prices

Intraday Prices

intraday(name)

intradays(...)

Arguments

name, ...
Company names as returned by the get_today() function

Details

Scraps the intraday stock prices for the provided companies

Returns a data.frame containing all the intraday prices for the specified stocks. Where the symbol column contains the name of the company.

Examples

head(intraday(), 30)

#>      labels   prices
#> 1  09:30:03 11033.09
#> 2  09:30:10 11033.27
#> 3  09:30:11 11034.06
#> 4  09:30:15 11028.71
#> 5  09:30:16 11030.31
#> 6  09:30:35 11037.64
#> 7  09:30:37 11037.62
#> 8  09:30:39 11037.64
#> 9  09:30:39 11036.13
#> 10 09:30:39 11035.81
#> 11 09:30:39 11054.57
#> 12 09:30:45 11054.28
#> 13 09:30:46 11055.77
#> 14 09:30:47 11045.97
#> 15 09:30:48 11045.99
#> 16 09:30:51 11045.64
#> 17 09:30:51 11045.67
#> 18 09:30:54 11045.62
#> 19 09:30:54 11045.83
#> 20 09:30:55 11046.21
#> 21 09:30:55 11047.40
#> 22 09:31:41 11049.99
#> 23 09:32:23 11049.66
#> 24 09:35:44 11049.20
#> 25 09:38:47 11045.30
#> 26 09:40:16 11042.62
#> 27 09:40:39 11043.07
#> 28 09:42:59 11043.08
#> 29 09:43:31 11043.41
#> 30 09:43:56 11041.46

intradays("Wafa Assur", "Nexans Maroc")

#>     labels  prices       symbol
#> 1 14:30:30  129.90 Nexans Maroc
#> 2 12:36:17 3661.00   Wafa Assur
#> 3 12:36:17 3661.00   Wafa Assur
#> 4 12:36:17 3660.00   Wafa Assur
#> 5 15:30:03 3475.00   Wafa Assur