Stock prices

Scraps Le Boursier's and gets the historical prices for the specified companies.

stock(x)

stocks(...)

Arguments

x, ...
Company name as returned by the get_today() function

Value

Returns a data.frame containing all the historical data for the specified stock.

date: the date of the given record.

open: the opening price of the stock on date.

high: the highest price.

low: the lowest price.

close: the closing price.

volume: Volume of stocks traded on date.

If the plural variant is used it'll add a symbol column that'll contain the name of the company.

Examples

head(stock("Wafa Assur"), 20)

#>          date open high  low close volume
#> 1  2015-12-02 3500 3500 3450  3450    332
#> 2  2015-12-03 3450 3452 3450  3450     84
#> 3  2015-12-04 3500 3500 3500  3500     10
#> 4  2015-12-07 3400 3400 3313  3313     56
#> 5  2015-12-09 3340 3340 3301  3301    169
#> 6  2015-12-10 3104 3151 3104  3151    633
#> 7  2015-12-11 3121 3340 3120  3340    304
#> 8  2015-12-15 3152 3165 3140  3165    155
#> 9  2015-12-16 3298 3298 3298  3298      7
#> 10 2015-12-18 3250 3250 3250  3250     27
#> 11 2015-12-21 3056 3275 3056  3275   2084
#> 12 2015-12-23 3259 3259 3250  3258   1155
#> 13 2015-12-28 3100 3279 3100  3279   1170
#> 14 2015-12-29 3157 3249 3126  3126   1096
#> 15 2015-12-30 2980 3279 2980  3279   2018
#> 16 2015-12-31 3270 3270 3270  3270    600
#> 17 2016-01-05 3130 3130 3130  3130    931
#> 18 2016-01-06 3131 3131 3130  3130    110
#> 19 2016-01-07 3300 3300 3300  3300      8
#> 20 2016-01-08 3280 3280 3280  3280   1000

head(stock(), 20) # same as runing stock("MASI")

#>          date    open    high     low   close volume
#> 1  2015-12-02 9082.84 9116.88 9050.97 9050.97     NA
#> 2  2015-12-03 9039.33 9064.14 9034.78 9052.79     NA
#> 3  2015-12-04 9075.98 9099.75 9027.76 9069.55     NA
#> 4  2015-12-07 9066.51 9068.24 9017.34 9017.34     NA
#> 5  2015-12-08 9032.88 9043.73 8998.44 9008.41     NA
#> 6  2015-12-09 8976.44 8990.56 8910.13 8910.13     NA
#> 7  2015-12-10 8912.39 8920.21 8858.52 8874.88     NA
#> 8  2015-12-11 8866.93 8883.76 8847.28 8856.28     NA
#> 9  2015-12-14 8871.15 8883.68 8836.80 8865.63     NA
#> 10 2015-12-15 8864.36 8884.87 8796.65 8796.65     NA
#> 11 2015-12-16 8796.65 8870.44 8796.65 8864.75     NA
#> 12 2015-12-17 8863.02 8870.07 8812.57 8846.45     NA
#> 13 2015-12-18 8843.70 8860.92 8821.78 8860.92     NA
#> 14 2015-12-21 8861.94 8864.25 8789.89 8864.25     NA
#> 15 2015-12-22 8862.78 8878.53 8848.41 8863.38     NA
#> 16 2015-12-23 8858.99 8893.10 8836.15 8893.10     NA
#> 17 2015-12-28 8885.18 8905.83 8836.22 8856.06     NA
#> 18 2015-12-29 8817.89 8873.96 8795.62 8873.96     NA
#> 19 2015-12-30 8873.72 8873.72 8802.90 8868.85     NA
#> 20 2015-12-31 8867.46 8925.71 8843.28 8925.71     NA

head(stock("MADEX"), 20)

#>          date    open    high     low   close volume
#> 1  2015-12-02 7412.29 7439.27 7381.59 7381.59     NA
#> 2  2015-12-03 7371.26 7393.34 7367.26 7383.32     NA
#> 3  2015-12-04 7402.89 7424.14 7363.78 7396.48     NA
#> 4  2015-12-07 7393.78 7395.28 7342.49 7342.49     NA
#> 5  2015-12-08 7355.33 7364.37 7326.03 7334.30     NA
#> 6  2015-12-09 7305.91 7318.45 7247.39 7247.39     NA
#> 7  2015-12-10 7249.40 7256.34 7204.57 7222.72     NA
#> 8  2015-12-11 7215.67 7230.61 7201.66 7208.81     NA
#> 9  2015-12-14 7222.01 7234.24 7192.55 7218.15     NA
#> 10 2015-12-15 7217.02 7235.38 7157.31 7157.31     NA
#> 11 2015-12-16 7157.31 7223.58 7157.31 7218.21     NA
#> 12 2015-12-17 7216.68 7222.93 7168.74 7198.81     NA
#> 13 2015-12-18 7196.37 7211.22 7178.34 7211.22     NA
#> 14 2015-12-21 7212.14 7217.25 7151.24 7217.25     NA
#> 15 2015-12-22 7216.69 7229.93 7202.33 7214.47     NA
#> 16 2015-12-23 7210.58 7239.25 7190.30 7239.25     NA
#> 17 2015-12-28 7232.22 7250.55 7188.78 7210.59     NA
#> 18 2015-12-29 7176.72 7223.24 7160.20 7223.24     NA
#> 19 2015-12-30 7223.02 7223.02 7161.35 7215.13     NA
#> 20 2015-12-31 7213.88 7255.21 7195.10 7255.21     NA

head(stocks("Masi", "MADEX"), 20)

#>          date    open    high     low   close volume symbol
#> 1  2015-12-02 9082.84 9116.88 9050.97 9050.97     NA   MASI
#> 2  2015-12-03 9039.33 9064.14 9034.78 9052.79     NA   MASI
#> 3  2015-12-04 9075.98 9099.75 9027.76 9069.55     NA   MASI
#> 4  2015-12-07 9066.51 9068.24 9017.34 9017.34     NA   MASI
#> 5  2015-12-08 9032.88 9043.73 8998.44 9008.41     NA   MASI
#> 6  2015-12-09 8976.44 8990.56 8910.13 8910.13     NA   MASI
#> 7  2015-12-10 8912.39 8920.21 8858.52 8874.88     NA   MASI
#> 8  2015-12-11 8866.93 8883.76 8847.28 8856.28     NA   MASI
#> 9  2015-12-14 8871.15 8883.68 8836.80 8865.63     NA   MASI
#> 10 2015-12-15 8864.36 8884.87 8796.65 8796.65     NA   MASI
#> 11 2015-12-16 8796.65 8870.44 8796.65 8864.75     NA   MASI
#> 12 2015-12-17 8863.02 8870.07 8812.57 8846.45     NA   MASI
#> 13 2015-12-18 8843.70 8860.92 8821.78 8860.92     NA   MASI
#> 14 2015-12-21 8861.94 8864.25 8789.89 8864.25     NA   MASI
#> 15 2015-12-22 8862.78 8878.53 8848.41 8863.38     NA   MASI
#> 16 2015-12-23 8858.99 8893.10 8836.15 8893.10     NA   MASI
#> 17 2015-12-28 8885.18 8905.83 8836.22 8856.06     NA   MASI
#> 18 2015-12-29 8817.89 8873.96 8795.62 8873.96     NA   MASI
#> 19 2015-12-30 8873.72 8873.72 8802.90 8868.85     NA   MASI
#> 20 2015-12-31 8867.46 8925.71 8843.28 8925.71     NA   MASI