Stock prices

Scraps Le Boursier's API and get the historical prices for the specified companies.

Returns a data.frame containing all the historical data for the specified stock. If the plural variant is used it'll add a symbol column that'll contain the name of the company.

stock(x)

stocks(...)

Arguments

x, ...
Company name as returned by the get_today() function

Details

Scraps the historical stock prices for the provided companies.

Examples

head(stock("Wafa Assur"), 20)

#>          date open high  low close volume
#> 1  2015-11-26 3690 3690 3690  3690      6
#> 2  2015-11-30 3580 3580 3500  3500    195
#> 3  2015-12-02 3500 3500 3450  3450    332
#> 4  2015-12-03 3450 3452 3450  3450     84
#> 5  2015-12-04 3500 3500 3500  3500     10
#> 6  2015-12-07 3400 3400 3313  3313     56
#> 7  2015-12-09 3340 3340 3301  3301    169
#> 8  2015-12-10 3104 3151 3104  3151    633
#> 9  2015-12-11 3121 3340 3120  3340    304
#> 10 2015-12-15 3152 3165 3140  3165    155
#> 11 2015-12-16 3298 3298 3298  3298      7
#> 12 2015-12-18 3250 3250 3250  3250     27
#> 13 2015-12-21 3056 3275 3056  3275   2084
#> 14 2015-12-23 3259 3259 3250  3258   1155
#> 15 2015-12-28 3100 3279 3100  3279   1170
#> 16 2015-12-29 3157 3249 3126  3126   1096
#> 17 2015-12-30 2980 3279 2980  3279   2018
#> 18 2015-12-31 3270 3270 3270  3270    600
#> 19 2016-01-05 3130 3130 3130  3130    931
#> 20 2016-01-06 3131 3131 3130  3130    110

head(stock(), 20) # same as runing stock("MASI")

#>          date    open    high     low   close volume
#> 1  2015-11-25 9103.71 9122.70 9095.00 9122.70     NA
#> 2  2015-11-26 9105.24 9119.59 9044.11 9108.93     NA
#> 3  2015-11-27 9108.71 9135.61 9079.33 9085.25     NA
#> 4  2015-11-30 9075.21 9113.27 9051.23 9093.21     NA
#> 5  2015-12-01 9085.12 9087.08 9050.43 9081.54     NA
#> 6  2015-12-02 9082.84 9116.88 9050.97 9050.97     NA
#> 7  2015-12-03 9039.33 9064.14 9034.78 9052.79     NA
#> 8  2015-12-04 9075.98 9099.75 9027.76 9069.55     NA
#> 9  2015-12-07 9066.51 9068.24 9017.34 9017.34     NA
#> 10 2015-12-08 9032.88 9043.73 8998.44 9008.41     NA
#> 11 2015-12-09 8976.44 8990.56 8910.13 8910.13     NA
#> 12 2015-12-10 8912.39 8920.21 8858.52 8874.88     NA
#> 13 2015-12-11 8866.93 8883.76 8847.28 8856.28     NA
#> 14 2015-12-14 8871.15 8883.68 8836.80 8865.63     NA
#> 15 2015-12-15 8864.36 8884.87 8796.65 8796.65     NA
#> 16 2015-12-16 8796.65 8870.44 8796.65 8864.75     NA
#> 17 2015-12-17 8863.02 8870.07 8812.57 8846.45     NA
#> 18 2015-12-18 8843.70 8860.92 8821.78 8860.92     NA
#> 19 2015-12-21 8861.94 8864.25 8789.89 8864.25     NA
#> 20 2015-12-22 8862.78 8878.53 8848.41 8863.38     NA

head(stock("MADEX"), 20)

#>          date    open    high     low   close volume
#> 1  2015-11-25 7420.81 7437.77 7413.07 7437.77     NA
#> 2  2015-11-26 7421.85 7435.50 7368.80 7435.50     NA
#> 3  2015-11-27 7435.61 7458.76 7413.26 7413.26     NA
#> 4  2015-11-30 7404.35 7438.40 7383.71 7421.15     NA
#> 5  2015-12-01 7414.07 7419.45 7383.39 7411.14     NA
#> 6  2015-12-02 7412.29 7439.27 7381.59 7381.59     NA
#> 7  2015-12-03 7371.26 7393.34 7367.26 7383.32     NA
#> 8  2015-12-04 7402.89 7424.14 7363.78 7396.48     NA
#> 9  2015-12-07 7393.78 7395.28 7342.49 7342.49     NA
#> 10 2015-12-08 7355.33 7364.37 7326.03 7334.30     NA
#> 11 2015-12-09 7305.91 7318.45 7247.39 7247.39     NA
#> 12 2015-12-10 7249.40 7256.34 7204.57 7222.72     NA
#> 13 2015-12-11 7215.67 7230.61 7201.66 7208.81     NA
#> 14 2015-12-14 7222.01 7234.24 7192.55 7218.15     NA
#> 15 2015-12-15 7217.02 7235.38 7157.31 7157.31     NA
#> 16 2015-12-16 7157.31 7223.58 7157.31 7218.21     NA
#> 17 2015-12-17 7216.68 7222.93 7168.74 7198.81     NA
#> 18 2015-12-18 7196.37 7211.22 7178.34 7211.22     NA
#> 19 2015-12-21 7212.14 7217.25 7151.24 7217.25     NA
#> 20 2015-12-22 7216.69 7229.93 7202.33 7214.47     NA

head(stocks("Masi", "MADEX"), 20)

#>          date    open    high     low   close volume symbol
#> 1  2015-11-25 9103.71 9122.70 9095.00 9122.70     NA   MASI
#> 2  2015-11-26 9105.24 9119.59 9044.11 9108.93     NA   MASI
#> 3  2015-11-27 9108.71 9135.61 9079.33 9085.25     NA   MASI
#> 4  2015-11-30 9075.21 9113.27 9051.23 9093.21     NA   MASI
#> 5  2015-12-01 9085.12 9087.08 9050.43 9081.54     NA   MASI
#> 6  2015-12-02 9082.84 9116.88 9050.97 9050.97     NA   MASI
#> 7  2015-12-03 9039.33 9064.14 9034.78 9052.79     NA   MASI
#> 8  2015-12-04 9075.98 9099.75 9027.76 9069.55     NA   MASI
#> 9  2015-12-07 9066.51 9068.24 9017.34 9017.34     NA   MASI
#> 10 2015-12-08 9032.88 9043.73 8998.44 9008.41     NA   MASI
#> 11 2015-12-09 8976.44 8990.56 8910.13 8910.13     NA   MASI
#> 12 2015-12-10 8912.39 8920.21 8858.52 8874.88     NA   MASI
#> 13 2015-12-11 8866.93 8883.76 8847.28 8856.28     NA   MASI
#> 14 2015-12-14 8871.15 8883.68 8836.80 8865.63     NA   MASI
#> 15 2015-12-15 8864.36 8884.87 8796.65 8796.65     NA   MASI
#> 16 2015-12-16 8796.65 8870.44 8796.65 8864.75     NA   MASI
#> 17 2015-12-17 8863.02 8870.07 8812.57 8846.45     NA   MASI
#> 18 2015-12-18 8843.70 8860.92 8821.78 8860.92     NA   MASI
#> 19 2015-12-21 8861.94 8864.25 8789.89 8864.25     NA   MASI
#> 20 2015-12-22 8862.78 8878.53 8848.41 8863.38     NA   MASI